refactor: Align ymax test/prod network definition file location and import patterns
